Move Modal and ConfirmModal components into `lego-bricks`
