+++ app/routes/admin/groups/components/GroupPermissions.tsx
@@ -1,10 +1,9 @@
-import { Flex, Icon } from '@webkom/lego-bricks';
+import { ConfirmModal, Flex, Icon } from '@webkom/lego-bricks';
 import { sortBy } from 'lodash';
 import { connect } from 'react-redux';
 import { Link } from 'react-router-dom';
 import { compose } from 'redux';
 import { editGroup } from 'app/actions/GroupActions';
-import { ConfirmModal } from 'app/components/Modal/ConfirmModal';
 import type { ID } from 'app/models';
 import loadingIndicator from 'app/utils/loadingIndicator';
 import AddGroupPermission from './AddGroupPermission';
