+++ app/routes/events/components/EventAdministrate/Attendees.tsx
@@ -1,8 +1,12 @@
-import { Button, Flex, LoadingIndicator } from '@webkom/lego-bricks';
+import {
+  Button,
+  ConfirmModal,
+  Flex,
+  LoadingIndicator,
+} from '@webkom/lego-bricks';
 import moment from 'moment-timezone';
 import { useState } from 'react';
 import { formatPhoneNumber, parsePhoneNumber } from 'react-phone-number-input';
-import { ConfirmModal } from 'app/components/Modal/ConfirmModal';
 import type {
   EventAdministrate,
   EventPool,
