+++ app/routes/photos/components/GalleryPictureModal.tsx
@@ -1,4 +1,4 @@
-import { Flex, Icon } from '@webkom/lego-bricks';
+import { Flex, Icon, Modal } from '@webkom/lego-bricks';
 import throttle from 'lodash/throttle';
 import { Component, useEffect } from 'react';
 import { Link } from 'react-router-dom';
@@ -7,7 +7,6 @@ import CommentView from 'app/components/Comments/CommentView';
 import { Content } from 'app/components/Content';
 import Dropdown from 'app/components/Dropdown';
 import { Image } from 'app/components/Image';
-import Modal from 'app/components/Modal';
 import ProgressiveImage from 'app/components/ProgressiveImage';
 import type { ID } from 'app/models';
 import type { EntityID } from 'app/types';
