+++ app/routes/users/components/Penalties.tsx
@@ -1,5 +1,4 @@
-import { Button } from '@webkom/lego-bricks';
-import { ConfirmModal } from 'app/components/Modal/ConfirmModal';
+import { Button, ConfirmModal } from '@webkom/lego-bricks';
 import { FormatTime } from 'app/components/Time';
 import type { Penalty } from 'app/models';
 import styles from './Penalties.css';
