+++ app/routes/users/components/UserProfile.tsx
@@ -1,4 +1,4 @@
-import { Button, Card, Flex, Icon } from '@webkom/lego-bricks';
+import { Button, Card, Flex, Icon, Modal } from '@webkom/lego-bricks';
 import cx from 'classnames';
 import { sumBy, sortBy, uniqBy, groupBy, orderBy } from 'lodash';
 import moment from 'moment-timezone';
@@ -9,7 +9,6 @@ import { Link } from 'react-router-dom';
 import frame from 'app/assets/frame.png';
 import EventListCompact from 'app/components/EventListCompact';
 import { ProfilePicture, CircularPicture, Image } from 'app/components/Image';
-import Modal from 'app/components/Modal';
 import Pill from 'app/components/Pill';
 import Tooltip from 'app/components/Tooltip';
 import type {
