+++ packages/lego-bricks/src/components/Modal/ConfirmModal.tsx
@@ -1,7 +1,9 @@
-import { Button, Flex, Icon } from '@webkom/lego-bricks';
 import { useState } from 'react';
-import Modal from 'app/components/Modal';
-import styles from './ConfirmModal.css';
+import { Button } from '../Button';
+import { Icon } from '../Icon';
+import { Flex } from '../Layout';
+import styles from './ConfirmModal.module.css';
+import { Modal } from './index';
 import type { ReactNode } from 'react';
 
 type ConfirmModalContentProps = {
@@ -69,7 +71,7 @@ type ConfirmModalProps = {
   danger?: boolean;
 };
 
-export const ConfirmModal = ({
+const ConfirmModal = ({
   onConfirm = async () => {},
   onCancel = async () => {},
   closeOnCancel = true,
@@ -97,7 +99,8 @@ export const ConfirmModal = ({
       } catch (error) {
         setWorking(false);
         setErrorMessage(
-          (error as any)?.meta?.errorMessage || 'Det skjedde en feil...'
+          (error as { meta?: { errorMessage: string } })?.meta?.errorMessage ||
+            'Det skjedde en feil...'
         );
         throw error;
       }
@@ -126,3 +129,5 @@ export const ConfirmModal = ({
     </>
   );
 };
+
+export default ConfirmModal;
