+++ packages/lego-bricks/src/components/Modal/index.ts
@@ -1,0 +1,2 @@
+export { default as Modal } from './Modal';
+export { default as ConfirmModal } from './ConfirmModal';
